Separate regression model for clarity
