+++ course_1_regression_and_classification/regression/main.py
@@ -1,6 +1,6 @@
 from course_1_regression_and_classification.regression.dataset_processing.dataset_utils import prepare_datasets
 from course_1_regression_and_classification.regression.optimization_by_regression import OptimizationByRegression
-from course_1_regression_and_classification.regression.regression_interface import Regression, Regressions
+from course_1_regression_and_classification.regression.regression_model.regression_interface import Regression, Regressions
 
 if __name__ == '__main__':
     train, validate, test = prepare_datasets("../../datasets/coursera_lin_reg_dataset.csv", delimiter=",")
