+++ course_1_regression_and_classification/regression/regression_model/regression_interface.py
@@ -1,6 +1,6 @@
 from enum import Enum
 
-from course_1_regression_and_classification.regression.abstract_regression import AbstractRegression
+from course_1_regression_and_classification.regression.regression_model.abstract_regression import AbstractRegression
 from course_1_regression_and_classification.regression.linear_regression.linear_reg import LinearRegression
 from course_1_regression_and_classification.regression.logistic_regression.logistic_reg import LogisticRegression
